Extract CommentSource class
